+++ xc3_model/src/map.rs
@@ -17,6 +17,7 @@ use xc3_lib::{
 use crate::{
     create_materials, create_samplers, lod_data, model_name,
     shader_database::{MapPrograms, ShaderDatabase},
+    skinning::create_skinning,
     texture::{self, CreateImageTextureError, ImageTexture},
     IndexMapExt, MapRoot, Material, Model, ModelBuffers, ModelGroup, Models, Texture,
 };
@@ -362,6 +363,7 @@ fn load_prop_model_group(
         models: Vec::new(),
         materials,
         samplers,
+        skinning: model_data.models.skinning.as_ref().map(create_skinning),
         lod_data: model_data.models.lod_data.as_ref().map(lod_data),
         morph_controller_names: Vec::new(),
         animation_morph_names: Vec::new(),
@@ -539,6 +541,7 @@ fn load_map_model_group(
         models,
         materials,
         samplers,
+        skinning: model_data.models.skinning.as_ref().map(create_skinning),
         lod_data: model_data.models.lod_data.as_ref().map(lod_data),
         morph_controller_names: Vec::new(),
         animation_morph_names: Vec::new(),
@@ -628,7 +631,8 @@ fn load_foliage_model(
                 models,
                 materials,
                 samplers: Vec::new(),
-                lod_data: None,
+                skinning: model_data.models.skinning.as_ref().map(create_skinning),
+                lod_data: model_data.models.lod_data.as_ref().map(lod_data),
                 morph_controller_names: Vec::new(),
                 animation_morph_names: Vec::new(),
                 min_xyz: model_data.models.min_xyz.into(),